refactor (ui): move types around
